requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
candlelite==1.0.9
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cryptography==41.0.5
Django==3.2
django-cors-headers==4.3.1
django-db-connection-pool==1.2.4
django-filter==2.4.0
django-redis==5.0.0
djangorestframework==3.12.4
greenlet==3.0.1
idna==3.4
loguru==0.7.2
numpy==1.26.1
okx==2.1.0
pandas==2.1.2
paux==1.0.14
pendulum==2.1.2
pycparser==2.21

PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
pytzdata==2020.1
redis==5.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.22
sqlparams==5.1.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.12

